\name{TDspatstat-package}
\alias{TDspatstat-package}
\alias{TDspatstat}
\docType{package}
\title{
\packageTitle{TDspatstat}
}
\description{
\packageDescription{TDspatstat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TDspatstat}
\packageIndices{TDspatstat}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TDspatstat}

Maintainer: \packageMaintainer{TDspatstat}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
